payloads sync across the network!
note: currently for paylods, we use the unity network transform component, instead of the client side prediction / snapshot process that is being used for the movement that is creating a slight lerp desync between the payload and the "tanks". additionally, player teams are determined by the modulus of the client index, for now.